src/requirements.txt
pandas==2.1.3
sqlalchemy==2.0.35
python-dotenv==1.0.1
yfinance==0.2.11
psycopg2-binary==2.9.10
